- [x] Create a service for isProductOwner - [x] Create a service for isShopOwner - [x] Add 404 code for non-existant ressources - [x] Use a transaction for add products to shop request - [x] Use camelCase for variable - [x] Use compact to pass variable to the front end - [x] Change the code for the verified midleware

+++ app/Http/Controllers/CategoryController.php
@@ -8,18 +8,19 @@ use App\Http\Requests\Category\CategorySearchRequest;
 use App\Http\Requests\Category\CategoryUpdateRequest;
 use App\Models\Category;
 use App\Models\Shop;
-use Exception;
+use App\Services\ProductOwnershipService;
+use App\Services\ShopOwnershipService;
 use Illuminate\Http\JsonResponse;
 
 class CategoryController extends Controller
 {
-    public function createCategory(CategoryRequest $request):JsonResponse
+    public function createCategory(CategoryRequest $request, ShopOwnershipService $shopOwnershipService):JsonResponse
     {
         $shop_id = $request->input('shop_id');
-        $shopController = new ShopController();
-        try {
-            $shopController->isShopOwner($shop_id);
-            $shop= Shop::find($shop_id);
+
+        if ($shopOwnershipService->isShopOwner($shop_id))
+        {
+            $shop= Shop::findOrFail($shop_id);
 
             $category=Category::create([
                 'name'=>$request->input('name'),
@@ -29,21 +30,17 @@ class CategoryController extends Controller
                 'message'=>'Category created successfully',
                 'category_id'=>$category->id,
             ],201);
-        } catch (Exception $e) {
-            return response()->json([
-                'message' => $e->getMessage()
-            ], 403);
         }
+        return response()->json([],403);
     }
 
-    public function searchCategory(CategorySearchRequest $request):JsonResponse
+    public function searchCategory(CategorySearchRequest $request,ShopOwnershipService $shopOwnershipService):JsonResponse
     {
         $shop_id=$request->input('shop_id');
         $search= $request->input('search');
 
-        $shopController = new ShopController();
-        try {
-            $shopController->isShopOwner($shop_id);
+        if ($shopOwnershipService->isShopOwner($shop_id))
+        {
             if ($search!=null)
             {
                 $category=Category::where('name','like',"%$search%")
@@ -58,82 +55,68 @@ class CategoryController extends Controller
             return response()->json([
                 'result'=>$category
             ]);
-        } catch (Exception $e) {
-            return response()->json([
-                'message' => $e->getMessage()
-            ], 403);
         }
+
+        return response()->json([],403);
     }
 
-    public function updateCategory(CategoryUpdateRequest $request,$id): JsonResponse
+    public function updateCategory(CategoryUpdateRequest $request,ShopOwnershipService $shopOwnershipService,$id): JsonResponse
     {
         $category = Category::find($id);
         if ($category!=null)
         {
             $shop_id=$category->shop_id;
-            $shopController = new ShopController();
-            try {
-                $shopController->isShopOwner($shop_id);
+            if ($shopOwnershipService->isShopOwner($shop_id))
+            {
                 $category->name=$request->input('name');
                 $category->save();
                 return response()->json([
                     'message'=>'Category updated successfully',
                     'category_id'=>$category->id,
                 ]);
-            } catch (Exception $e) {
-                return response()->json([
-                    'message' => $e->getMessage()
-                ], 403);
             }
+            return response()->json([],403);
         }
 
         return response()->json([
             'message'=>'The category does not exist'
-        ]);
+        ],404);
     }
 
-    public function deleteCategory($id): JsonResponse
+    public function deleteCategory(ShopOwnershipService $shopOwnershipService,$id): JsonResponse
     {
         $category= Category::find($id);
         if ($category!=null)
         {
             $shop_id=$category->shop_id;
-            $shopController = new ShopController();
-            try {
-                $shopController->isShopOwner($shop_id);
+            if ($shopOwnershipService->isShopOwner($shop_id))
+            {
                 $category->products()->detach();
                 $category->delete();
                 return response()->json([],204);
-            } catch (Exception $e) {
-                return response()->json([
-                    'message' => $e->getMessage()
-                ], 403);
             }
+            return response()->json([],404);
         }
         return response()->json([
             'message'=>'The category does not exist'
-        ]);
+        ],404);
     }
 
-    public function getCategoryProducts($id): JsonResponse
+    public function getCategoryProducts(ShopOwnershipService $shopOwnershipService,$id): JsonResponse
     {
         $category=Category::find($id);
 
         if ($category!=null)
         {
-            $shopController = new ShopController();
-            try {
-                $shopController->isShopOwner($category->shop_id);
+            if ($shopOwnershipService->isShopOwner($category->shop_id))
+            {
                 $products=$category->products()->paginate(15);
 
                 return response()->json([
                     'products'=>$products,
                 ]);
-            } catch (Exception $e) {
-                return response()->json([
-                    'message' => $e->getMessage()
-                ], 403);
             }
+            return response()->json([],403);
         }
 
         return response()->json([
@@ -141,63 +124,58 @@ class CategoryController extends Controller
         ],404);
     }
 
-    public function addProductsToCategory(AddToCategoryRequest $request): JsonResponse
+    public function addProductsToCategory(AddToCategoryRequest $request, ProductOwnershipService $productOwnershipService,ShopOwnershipService $shopOwnershipService): JsonResponse
     {
         $category = Category::find($request->input('category_id'));
         if ($category !== null) {
-            $shopController = new ShopController();
-            $productController = new ProductController();
-            try {
-                $shopController->isShopOwner($category->shop_id);
+                if ($shopOwnershipService->isShopOwner($category->shop_id))
+                {
+                    $productsId = $request->input('products');
+                    foreach ($productsId as $productId) {
+                        if (!$productOwnershipService->isProductOwner($productId))
+                        {
+                            return response()->json([],403);
+                        }
+                    }
+                    $category->products()->attach($productsId);
 
-                $productsId = $request->input('products');
-                foreach ($productsId as $productId) {
-                    $productController->isProductOwner($productId);
+                    return response()->json([
+                        'message' => 'Products added successfully'
+                    ]);
                 }
-                $category->products()->attach($productsId);
-
-                return response()->json([
-                    'message' => 'Products added successfully'
-                ]);
-            } catch (\Exception $e) {
-                return response()->json([
-                    'message' => $e->getMessage()
-                ], 403);
-            }
+                return response()->json([],403);
         }
         return response()->json([
             'message' => 'The category does not exist',
-        ]);
+        ],404);
     }
 
 
-    public function removeProductsFromCategory(AddToCategoryRequest $request): JsonResponse
+    public function removeProductsFromCategory(AddToCategoryRequest $request,ProductOwnershipService $productOwnershipService, ShopOwnershipService $shopOwnershipService): JsonResponse
     {
         $category = Category::find($request->input('category_id'));
         if ($category!=null)
         {
-            $shopController = new ShopController();
-            $productController = new ProductController();
-            try {
-                $shopController->isShopOwner($category->shop_id);
-
+            if ($shopOwnershipService->isShopOwner($category->shop_id))
+            {
                 $productsId = $request->input('products');
                 foreach ($productsId as $productId) {
-                    $productController->isProductOwner($productId);
+                    if (!$productOwnershipService->isProductOwner($productId))
+                    {
+                        return response()->json([],403);
+                    }
                 }
                 $category->products()->detach($productsId);
 
                 return response()->json([
                     'message'=>'Products successfully removed form the category',
                 ]);
-            } catch (Exception $e) {
-                return response()->json([
-                    'message' => $e->getMessage()
-                ], 403);
             }
+            return  response()->json([],403);
+
         }
         return response()->json([
             'message'=>'The category does not exist',
-        ]);
+        ],404);
     }
 }
